requirements.txt
torch>=2.0.0
snowflake-snowpark-python>=1.11.0
snowflake-connector-python>=3.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
